Change order of inheritance
